// frontend/src/app/models/userModel.ts
export class UserModel {
    id: number;
    username: string;
    email: string;
    firstName: string;
    lastName: string;
    birthDate: string;
    city: string;
    phoneNumber: string;
    agency: string;
    agencyAddress: string;
    agencyCity: string;
    agencyPIB: string;
    licenseNo: number;
    userType: string;
    waitingApproval: boolean;
    denied: boolean;
}